+++ src/Exercism.Analyzers.CSharp/Analyzers/Default/DefaultAnalyzer.cs
@@ -1,0 +1,6 @@
+namespace Exercism.Analyzers.CSharp.Analyzers.Default;
+
+internal class DefaultAnalyzer : ExerciseAnalyzer<DefaultSolution>
+{
+    protected override SolutionAnalysis AnalyzeSpecific(DefaultSolution solution) => solution.Analysis;
+}
