--- src/Exercism.Analyzers.CSharp/Analyzers/Default/DefaultExerciseAnalyzer.cs
@@ -1,8 +1,0 @@
-using Exercism.Analyzers.CSharp.Analyzers.Shared;
-
-namespace Exercism.Analyzers.CSharp.Analyzers.Default;
-
-internal class DefaultExerciseAnalyzer : SharedAnalyzer<DefaultSolution>
-{
-    protected override SolutionAnalysis AnalyzeSpecific(DefaultSolution solution) => null;
-}
