+++ src/Exercism.Analyzers.CSharp/Analyzers/Leap/LeapAnalyzer.cs
@@ -1,17 +1,12 @@
-using Exercism.Analyzers.CSharp.Analyzers.Shared;
-
 using static Exercism.Analyzers.CSharp.Analyzers.Leap.LeapComments;
 using static Exercism.Analyzers.CSharp.Analyzers.Leap.LeapSolution;
 using static Exercism.Analyzers.CSharp.Analyzers.Shared.SharedComments;
 
 namespace Exercism.Analyzers.CSharp.Analyzers.Leap;
 
-internal class LeapAnalyzer : SharedAnalyzer<LeapSolution>
+internal class LeapAnalyzer : ExerciseAnalyzer<LeapSolution>
 {
-    protected override SolutionAnalysis AnalyzeSpecific(LeapSolution solution) =>
-        DisapproveWhenInvalid(solution) ?? ApproveWhenValid(solution);
-    
-    private SolutionAnalysis DisapproveWhenInvalid(LeapSolution solution)
+    protected override SolutionAnalysis AnalyzeSpecific(LeapSolution solution)
     {
         if (solution.MissingLeapClass)
             solution.AddComment(MissingClass(LeapClassName));
@@ -31,23 +26,12 @@ internal class LeapAnalyzer : SharedAnalyzer<LeapSolution>
         if (solution.UsesTooManyChecks)
             solution.AddComment(UseMinimumNumberOfChecks);
 
-        return solution.HasComments
-            ? solution.Disapprove()
-            : solution.ContinueAnalysis();
-    }
-
-    private SolutionAnalysis ApproveWhenValid(LeapSolution solution)
-    {
         if (solution.UsesIfStatement)
             solution.AddComment(DoNotUseIfStatement);
 
         if (solution.UsesSingleLine && !solution.UsesExpressionBody)
             solution.AddComment(UseExpressionBodiedMember(IsLeapYearMethodName));
 
-        if (solution.ReturnsMinimumNumberOfChecksInSingleExpression ||
-            solution.HasComments)
-            return solution.Approve();
-
-        return solution.ContinueAnalysis();
+        return solution.Analysis;
     }
 }
