+++ src/Exercism.Analyzers.CSharp/Analyzers/TwoFer/TwoFerAnalyzer.cs
@@ -1,17 +1,12 @@
-using Exercism.Analyzers.CSharp.Analyzers.Shared;
-
 using static Exercism.Analyzers.CSharp.Analyzers.Shared.SharedComments;
 using static Exercism.Analyzers.CSharp.Analyzers.TwoFer.TwoFerComments;
 using static Exercism.Analyzers.CSharp.Analyzers.TwoFer.TwoFerSolution;
 
 namespace Exercism.Analyzers.CSharp.Analyzers.TwoFer;
 
-internal class TwoFerAnalyzer : SharedAnalyzer<TwoFerSolution>
+internal class TwoFerAnalyzer : ExerciseAnalyzer<TwoFerSolution>
 {
-    protected override SolutionAnalysis AnalyzeSpecific(TwoFerSolution solution) =>
-        DisapproveWhenInvalid(solution) ?? ApproveWhenValid(solution);
-    
-    private SolutionAnalysis DisapproveWhenInvalid(TwoFerSolution solution)
+    protected override SolutionAnalysis AnalyzeSpecific(TwoFerSolution solution)
     {
         if (solution.MissingTwoFerClass)
             solution.AddComment(MissingClass(TwoFerClassName));
@@ -43,13 +38,6 @@ internal class TwoFerAnalyzer : SharedAnalyzer<TwoFerSolution>
         if (solution.UsesStringConcat)
             solution.AddComment(UseStringInterpolationNotStringConcat);
 
-        return solution.HasComments
-            ? solution.Disapprove()
-            : solution.ContinueAnalysis();
-    }
-
-    private SolutionAnalysis ApproveWhenValid(TwoFerSolution solution)
-    {
         if (solution.UsesStringConcatenation)
             solution.AddComment(UseStringInterpolationNotStringConcatenation);
 
@@ -65,56 +53,15 @@ internal class TwoFerAnalyzer : SharedAnalyzer<TwoFerSolution>
         if (solution.UsesNullCheck)
             solution.AddComment(UseNullCoalescingOperatorNotNullCheck);
 
-        if (solution.UsesSingleLine)
-            return AnalyzeSingleLine(solution);
-
-        if (solution.AssignsToParameter)
-            return AnalyzeParameterAssignment(solution);
-
-        if (solution.AssignsVariable)
-            return AnalyzeVariableAssignment(solution);
-
-        return solution.ContinueAnalysis();
-    }
-
-    private static SolutionAnalysis AnalyzeSingleLine(TwoFerSolution solution)
-    {
         if (!solution.UsesExpressionBody)
             solution.AddComment(UseExpressionBodiedMember(SpeakMethodName));
-
-        if (solution.UsesStringInterpolationWithDefaultValue ||
-            solution.UsesStringInterpolationWithNullCoalescingOperator ||
-            solution.HasComments)
-            return solution.Approve();
-
-        return solution.ContinueAnalysis();
-    }
-
-    private static SolutionAnalysis AnalyzeParameterAssignment(TwoFerSolution solution)
-    {
+        
         if (!solution.AssignsParameterUsingKnownExpression)
-            return solution.ReferToMentor();
+            return solution.Analysis;
 
         if (solution.AssignsParameterUsingNullCoalescingOperator)
             solution.AddComment(DoNotAssignAndReturn);
 
-        if (solution.ReturnsStringInterpolation ||
-            solution.HasComments)
-            return solution.Approve();
-
-        return solution.ContinueAnalysis();
-    }
-
-    private static SolutionAnalysis AnalyzeVariableAssignment(TwoFerSolution solution)
-    {
-        if (!solution.AssignsVariableUsingKnownInitializer)
-            return solution.ReferToMentor();
-
-        if (solution.ReturnsStringInterpolationWithVariable &&
-            solution.AssignsVariableUsingNullCoalescingOperator ||
-            solution.HasComments)
-            return solution.Approve();
-
-        return solution.ContinueAnalysis();
+        return solution.Analysis;
     }
 }
